New class to interact with a Wiimote
The class uses the cwiid library from http://abstrakraft.org/cwiid/
to detect the wiimote, read its settings and receive events.

Currently supported:
 - buttons
 - switching LEDs on and off
 - rumble to switch haptic feedback on and off
 - accelerometers: x, y and z

There's a small demo application called wiimote-demo that can be used
as an example how to use this class, or to just play around with the
device. When starting wiimote-demo, press the 1 and 2 buttons on your
wiimote and it will connect to it. This happens in two steps, first
the address of the wiimote needs to be found, then the actual
connection is established.

The class is async, it does interaction with the wiimote (which is
often blocking) in its own thread, sending over information using
signals.

Have fun :)

CCMAIL:plasma-devel@kde.org

svn path=/trunk/playground/base/plasma/MediaCenterComponents/; revision=1109276