Flick scrolling using the A button
This needs a patch in kdelibs, which I'll also commit. The kdelibs change
basically exposes flicking in both directions in the ScrollWidget API.
We'll see if this can go in like that, or if we should change it, it
probably would be nice to make it a slot.

Also add a bit of UI to make the plasmoid more self-explanatory.

svn path=/trunk/playground/base/plasma/MediaCenterComponents/; revision=1111196