Revert "Do not find both QxOrm release and debug library"
This reverts commit d984019c1e3ab4433ff85f2650c65a45178f5e62.

This change wrongly assumes that distributions build packages in debug mode
(which is incorrect in most cases).

The correct solution is to either recommend using qxorm 1.2.5 or patch 1.2.6 using
Shantanu's patch.

Shantanu, I could not find a git/svn/whatever repository for QxOrm, can your patch be added
in the pmc repo and the doc updated to mention it?

CCMAIL: shantanu@kde.org