Delay before starting media sources
The MediaLibrary waits for some time before emitting newMedia, newAlbum
etc so that it can batch multiple items together. Due to this, we have
to make sure that the first time we populate the library, we give
enough time to consumers to update themselves before we start possibly
resource hungry media sources.