Do not try to guess title for audio before getting metadata
We use Baloo in two stages, first stage we get only basic file details
which includes the filename. The second stage we get metadata about
audio files. It causes unncessary updates if we guess a title in the
first go and then the real title in second go. So, its better to just
do it in the second go.