Do not cache media lists to a database
We started caching results from media sources due to our primary data
source (Nepomuk) was quite slow and the experience was quite bad both
for fetching media and searching through it. Going forward, Baloo is
our primary data source, and was found to be quite performant. Given
this, the additional complexity of maintaining a cache in DB is too much
for the 1-2 second increase in Media Library load times.

Note that we are still able to pull in data from multiple data sources,
we just won't store it on disk.

REVIEW: 117867