Fetch photo-taken Date/Time for Images and file created Date/Time for other files
Just like the Nepomuk media source, the Baloo media source needs to
provide the date/time information for media. This is used to sort the
media to show the more recent media first.
For images, the date/time when the photo was actually taken is used,
and the file creation date/time is used for other media.

REVIEW: 117915