Take care of thread affinity of objects and their parents
Objects in a thread should not be created dynamically in their ctors.
They should instead be created in a method that is called directly or
indirectly by invocation of QThread::run() [as a result of someone
calling start() on the thread].
Also, threads should be moved to their own threads so that they truly
run independent of other threads. (Why isn't this the default?)
Finally, threads should not be parented to an object in another thread.