Build filesystemsearch plugin even if the baloo is installed
Earlier we assumed that if baloo is available at build time it will also
be available at runtime but this is not case for packagers. They may
have baloo at build time but users may not have baloo at runtime causing
no media files to be shown as there are no mediasource plugins.
BUG: 346217
FIXED-IN: 5.3